Bugs After Migration
List of bugs after migration
The position of the ProjectListView for all project should be fixed at
the same level. Done
The index when DidSelectRow on the table for Project ListView is wrong
and displaying on edit mode the wrong project. Done
Change all icons to to SampleColor.jpg in project root. Done
Missing segue for gear button on ProjectListView (Parameters). Done
Data Collecting is crashing after importing a pdf from forms. Done
Replace the Infield logo on report for AirSite Done
Replace the splash logo with AirSite also wsp on bottom. Done